+++ CHANGELOG.rst
@@ -104,7 +104,7 @@ Other changes
 - Clean up local test-build version and entries.
 - Update `setup.cfg` comments (fix typo, succinctness, punctuation).
 - Add Wayland-related test-handling to `test_interface.py`.
-
+- Handle Window import error on scripting tests in `lib/autokey/scripting/__init__.py`.
 
 .. _its counterpart: https://github.com/autokey/autokey/blob/master/.github/ISSUE_TEMPLATE/config.yml